+++ toolium/behave/environment.py
@@ -34,19 +34,19 @@ def before_all(context):
     context.logger = logging.getLogger()
 
     # Get default driver wrapper
-    wrapper = DriverWrappersPool.get_default_wrapper()
+    context.driver_wrapper = DriverWrappersPool.get_default_wrapper()
 
     # Configure wrapper
     if not hasattr(context, 'config_files'):
         context.config_files = ConfigFiles()
-    wrapper.configure(True, context.config_files)
-    context.config = wrapper.config
+    context.driver_wrapper.configure(True, context.config_files)
+    context.toolium_config = context.driver_wrapper.config
 
     # Create driver if it must be reused
-    context.reuse_driver = wrapper.config.getboolean_optional('Common', 'reuse_driver')
+    context.reuse_driver = context.driver_wrapper.config.getboolean_optional('Common', 'reuse_driver')
     if context.reuse_driver:
-        context.driver = wrapper.connect()
-        context.utils = wrapper.utils
+        context.driver = context.driver_wrapper.connect()
+        context.utils = context.driver_wrapper.utils
 
 
 def before_scenario(context, scenario):
@@ -56,20 +56,20 @@ def before_scenario(context, scenario):
     :param scenario: running scenario
     """
     # Get default driver wrapper
-    driver_wrapper = DriverWrappersPool.get_default_wrapper()
+    context.driver_wrapper = DriverWrappersPool.get_default_wrapper()
 
     # Create driver if it must not be reused
     if not context.reuse_driver:
         # Configure wrapper
-        driver_wrapper.configure(True, context.config_files)
-        context.config = driver_wrapper.config
+        context.driver_wrapper.configure(True, context.config_files)
+        context.toolium_config = context.driver_wrapper.config
 
         # Create driver
-        context.driver = driver_wrapper.connect()
-        context.utils = driver_wrapper.utils
+        context.driver = context.driver_wrapper.connect()
+        context.utils = context.driver_wrapper.utils
 
     # Common initialization
-    bdd_common_before_scenario(context, scenario, driver_wrapper)
+    bdd_common_before_scenario(context, scenario, context.driver_wrapper)
 
 
 def bdd_common_before_scenario(context_or_world, scenario, driver_wrapper):
@@ -91,16 +91,12 @@ def bdd_common_before_scenario(context_or_world, scenario, driver_wrapper):
 
     context_or_world.assertScreenshot = assertScreenshot
     context_or_world.assertFullScreenshot = assertFullScreenshot
+    context_or_world.app_strings = driver_wrapper.app_strings
 
     # Add implicitly wait
     implicitly_wait = driver_wrapper.config.get_optional('Common', 'implicitly_wait')
     if implicitly_wait:
         context_or_world.driver.implicitly_wait(implicitly_wait)
-
-    # Get application strings
-    if driver_wrapper.is_mobile_test() and not driver_wrapper.is_web_test() and not hasattr(context_or_world,
-                                                                                            'app_strings'):
-        context_or_world.app_strings = driver_wrapper.driver.app_strings()
 
     context_or_world.logger.info("Running new scenario: {0}".format(scenario.name))