+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Toolium Changelog
 =================
 
+v0.12.1
+-------
+
+*Release date: 2016-01-07*
+
+- Fix app_strings initialization in Behave Appium tests
+- In Behave tests, Toolium config is saved in context.toolium_config instead of using context.config to avoid
+  overriding Behave config
+
 v0.12.0
 -------